Add .gitattribute file to recognize files as regex
This is for the benefit of GitHub to recognize the content files as
regular expressions and mark them as such. I had considered marking
the repository as being C#, since that was the intended usage of the
patterns, but opted for regular expression, since that is what the code
really is. While it might need modification to work in other languages,
it is basically language agnostic.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>