Adjust given language in .gitattributes
Switch the named language in .gitattributes. Tried to use the alias of
regex, as listed in the linguist libraries languages.yml file. Seems
that the alias is for some other use. Hoped I could use that since I am
not sure it the language should be quoted or not since it has a space
in the name. Trying it without, since it looks like single line
declarations and might be the proper manner.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>